% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin Choose}
\usage{
bin_choose(n = 0, k = 0)
}
\arguments{
\item{n}{trials}

\item{k}{successes}
}
\value{
a single value representing amount of possibilities for k successes
}
\description{
after putting in number of successes and trials, this outputs the number of times that a certain number of successes can occur.
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
