% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials = 0, prob = 0)
}
\arguments{
\item{trials}{the amount of trials the test is occuring}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
a table, with classes bincum and data.frame, with three columns, the number of successes, the probability of it occuring, and the cumulative probability as number of successes increases.
}
\description{
this takes into account trials and probability in order to relay the overal distribution of probability of success from 0 to the amount of trails, while also taking into account the cumulative amount of probability used.
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
