% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Bin Distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within one trial}
}
\value{
a table of the probability distribution (with classes bindis and data.frame) from 0 successes to the same amount of success as trials.
}
\description{
this takes into account trials and probability in order to relay the overall distribution of the probability of success from 0 to the amount of trials.
}
\examples{
bin_distribution(trials = 5, prob = .5)
dis1 <- bin_distribution(trials = 5, prob = 0.5)
plot.bindis(dis1)
}
