% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_mean}
\alias{bin_mean}
\title{Bin Mean}
\usage{
bin_mean(trials = 0, prob = 0)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
the value of mean given the above parameters
}
\description{
this function finds the mean given a tests number of trials and probability of success.
}
\examples{
bin_mean(10, 0.3)
}
