% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_mode}
\alias{bin_mode}
\title{Bin Mode}
\usage{
bin_mode(trials = 0, prob = 0)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
the value of mode given the above parameters
}
\description{
this function finds the mode given a tests number of trials and probability of success.
}
\examples{
bin_mode(10, 0.3)
}
