% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_probability}
\alias{bin_probability}
\title{Bin Probability}
\usage{
bin_probability(success = 0, trials = 0, prob = 0)
}
\arguments{
\item{success}{the amount of successes testing for}

\item{trials}{the amount of trials in the test}

\item{prob}{the probability that a given success will occur within one trial}
}
\value{
this returns a single value representing the probability of the given parameters inputted.
}
\description{
takes into account successes, trials, and probability in order to relay the probability of achieving a certain binomial distribution.
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
