% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{Bin Skewness}
\usage{
bin_skewness(trials = 0, prob = 0)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
the value of skewness given the above parameters
}
\description{
this function finds the skewness given a tests number of trials and probability of success.
}
\examples{
bin_skewness(10, 0.3)
}
