% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin Variable}
\usage{
bin_variable(trials = trials, prob = prob)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
a stringed list, with class binvar, of the number of trials and probability of success, while also having several methods in order to print a list of the summary data based on these parameters.
}
\description{
this takes into account trials and probability in order to output a list which states the number of trials and probability that you entered.
}
\examples{
bin1 <- bin_variable(trials = 10, p = 0.3)
binsum1 <- summary(bin1)
binsum1
}
