% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{bin_variance}
\alias{bin_variance}
\title{Bin Variance}
\usage{
bin_variance(trials = 0, prob = 0)
}
\arguments{
\item{trials}{the amount of trials being tested for}

\item{prob}{the probability that a success will occur within a trial}
}
\value{
the value of variance given the above parameters
}
\description{
this function finds the variance given a tests number of trials and probability of success.
}
\examples{
bin_variance(10, 0.3)
}
